refactor(dashboard): change directory to dashboard
