+++ backend/src/test/java/com/brokendev/backend/services/DashboardServiceTest.java
@@ -4,10 +4,11 @@ import com.brokendev.backend.account.service.AccountService;
 import com.brokendev.backend.common.domain.user.User;
 import com.brokendev.backend.account.domain.Account;
 import com.brokendev.backend.account.dto.TransactionStatementResponseDTO;
-import com.brokendev.backend.dto.dashboard.DashboardResponseDTO;
+import com.brokendev.backend.dashboard.dto.DashboardResponseDTO;
 import com.brokendev.backend.common.exceptions.UserAccountNotFoundException;
 import com.brokendev.backend.account.domain.AccountRepository;
 import com.brokendev.backend.common.domain.user.UserRepository;
+import com.brokendev.backend.dashboard.service.DashboardService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
